backend/app/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.25.0
python-dotenv>=0.19.0